requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
pdfplumber>=0.10.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
